60FPS: Fix Great Glacier counter steps before fainting
Also refactor MOVIE FPS fix by moving it all into ff7/field.cpp
