+++ src/movies.h
@@ -23,7 +23,8 @@
 
 #include "common.h"
 
-extern short movie_fps_ratio;
+extern short movie_fps_ratio = 1;
+extern bool is_movie_bgfield = false;
 
 bool is_overlapping_movie_playing();
 void movie_init();
